Created Libs for sharing DTO beetwen microservice
