+++ apps/core_app/src/application/commands/users_cases/create-user.use-case.ts
@@ -1,9 +1,7 @@
 import { CommandHandler } from "@nestjs/cqrs"
-import { User } from "@prisma/client"
 import { UsersRepository } from "apps/core_app/src/infrastructure/modules/users/repositories/user.repository"
-import { CreateUserDto } from "../../dto/CreateUserDto"
-import { IUserInterface } from "../../services/user/user-interface"
 import { UserViewModel } from "apps/core_app/src/domain/interfaces/view_models/UserViewModel"
+import { CreateUserDto } from "@app/shared-dto"
 
 
 export class CreateUserCommand {
