+++ apps/core_app/src/application/commands/users_cases/get-user-by-username.use-case.ts
@@ -1,8 +1,5 @@
 import { CommandHandler } from "@nestjs/cqrs"
-import { User } from "@prisma/client"
 import { UsersRepository } from "apps/core_app/src/infrastructure/modules/users/repositories/user.repository"
-import { CreateUserDto } from "../../dto/CreateUserDto"
-import { IUserInterface } from "../../services/user/user-interface"
 import { UserViewModel } from "apps/core_app/src/domain/interfaces/view_models/UserViewModel"
