+++ libs/shared-dto/src/index.ts
@@ -1,0 +1,4 @@
+export * from './shared-dto.module';
+export * from './shared-dto.service';
+export * from './dtos/create-user.dto';
+
