+++ libs/shared-dto/src/shared-dto.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { SharedDtoService } from './shared-dto.service';
+
+@Module({
+  providers: [SharedDtoService],
+  exports: [SharedDtoService],
+})
+export class SharedDtoModule {}
