+++ package.json
@@ -37,6 +37,7 @@
     "generate:auth": "npx prisma generate --schema=apps/auth/prisma/schema.prisma"
   },
   "dependencies": {
+    "@libs/shared-dto": "file:./libs/shared-dto",
     "@nestjs/axios": "^3.1.3",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.3.0",
@@ -102,8 +103,12 @@
     "coverageDirectory": "./coverage",
     "testEnvironment": "node",
     "roots": [
-      "<rootDir>/apps/"
-    ]
+      "<rootDir>/apps/",
+      "<rootDir>/libs/"
+    ],
+    "moduleNameMapper": {
+      "^@app/shared-dto(|/.*)$": "<rootDir>/libs/shared-dto/src/$1"
+    }
   },
   "main": "index.js",
   "repository": "https://github.com/Smart-team-It-incubator/backend-inctagram"
