+++ tsconfig.json
@@ -17,6 +17,13 @@
     "strictBindCallApply": false,
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
-    "paths": {}
+    "paths": {
+      "@app/shared-dto": [
+        "libs/shared-dto/src"
+      ],
+      "@app/shared-dto/*": [
+        "libs/shared-dto/src/*"
+      ]
+    }
   }
 }
